% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efmedexpr.R
\name{meditation_expertise}
\alias{meditation_expertise}
\title{Calculate Meditation Expertise}
\usage{
meditation_expertise(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame
}
\description{
Calculates meditation expertise in hours (Hasenkamp & Barsalou, 2012, p.11).  Expects a data frame
formatted by \code{\link[expfactoryr:process_expfactory_survey]{expfactoryr::process_expfactory_survey}}.
}
\examples{

expertise_df <- expand.grid(token = participants$token, survey = 'meditation-expertise-results.json') \%>\%
  rowwise() \%>\%
  do(., expfactoryr::process_expfactory_survey(.$token, paste('data/', .$token, '_finished/', .$survey, sep=''),
    flat=TRUE))
 expertise <- efmedexpr::meditation_expertise(expertise_df)

}
\keyword{meditation,expertise}
